[Views] split views.py into view files for each Model
