+++ cms/views/home_view.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def index(request):
+    return render(request, 'cms/index.html.haml')
